requirements.txt
boto3==1.23.10
botocore==1.26.10
dnspython==2.2.1
jmespath==0.10.0
joblib==1.1.1
numpy==1.19.5
pandas==1.1.5
pymongo==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.1
s3transfer==0.5.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.18
